+++ bin/check-rabbitmq-node-usage.rb
@@ -34,6 +34,7 @@
 
 require 'sensu-plugin/check/cli'
 require 'carrot-top'
+require 'inifile'
 
 # Main plugin class
 class CheckRabbitMQNodeUsage < Sensu::Plugin::Check::CLI
@@ -142,6 +143,11 @@ class CheckRabbitMQNodeUsage < Sensu::Plugin::Check::CLI
          long: '--dcrit PERCENT',
          proc: proc(&:to_f),
          default: 90
+
+  option :ini,
+         description: 'Configuration ini file',
+         short: '-i',
+         long: '--ini VALUE'
 
   def run
     res = node_healthy?
@@ -293,11 +299,21 @@ class CheckRabbitMQNodeUsage < Sensu::Plugin::Check::CLI
   end
 
   def rabbitmq_management
+    if config[:ini]
+      ini = IniFile.load(config[:ini])
+      section = ini['auth']
+      username = section['username']
+      password = section['password']
+    else
+      username = config[:username]
+      password = config[:password]
+    end
+
     CarrotTop.new(
       host: config[:host],
       port: config[:port],
-      user: config[:username],
-      password: config[:password],
+      user: username,
+      password: password,
       ssl: config[:ssl]
     )
   end