+++ bin/metrics-rabbitmq-queue.rb
@@ -29,6 +29,7 @@
 require 'sensu-plugin/metric/cli'
 require 'socket'
 require 'carrot-top'
+require 'inifile'
 
 # main plugin class
 class RabbitMQMetrics < Sensu::Plugin::Metric::CLI::Graphite
@@ -48,9 +49,9 @@ class RabbitMQMetrics < Sensu::Plugin::Metric::CLI::Graphite
          short: '-v',
          long: '--vhost VHOST'
 
-  option :user,
+  option :username,
          description: 'RabbitMQ management API user',
-         long: '--user USER',
+         long: '--username USER',
          default: 'guest'
 
   option :password,
@@ -73,13 +74,28 @@ class RabbitMQMetrics < Sensu::Plugin::Metric::CLI::Graphite
          boolean: true,
          default: false
 
+  option :ini,
+         description: 'Configuration ini file',
+         short: '-i',
+         long: '--ini VALUE'
+
   def acquire_rabbitmq_queues
     begin
+      if config[:ini]
+        ini = IniFile.load(config[:ini])
+        section = ini['auth']
+        username = section['username']
+        password = section['password']
+      else
+        username = config[:username]
+        password = config[:password]
+      end
+
       rabbitmq_info = CarrotTop.new(
         host: config[:host],
         port: config[:port],
-        user: config[:user],
-        password: config[:password],
+        user: username,
+        password: password,
         ssl: config[:ssl]
       )
     rescue